fix remarks 2
