+++ task3/test-students/src/main/java/ru/otus/teststudents/config/LocaleConfig.java
@@ -12,6 +12,4 @@ public interface LocaleConfig {
     Locale getCurrentLocale();
 
     void setCurrentLocale(Locale locale);
-
-    String getQuestionFileName();
 }
