+++ task3/test-students/src/main/java/ru/otus/teststudents/config/QuestionStorageConfig.java
@@ -1,0 +1,6 @@
+package ru.otus.teststudents.config;
+
+public interface QuestionStorageConfig {
+
+    String getQuestionFileName();
+}
